src/tasks/vision/03-zero-shot-object-detection/requirements.txt
transformers
torch
pillow
scipy
